refactor: worked on the suggested changes
