+++ src/context/context.jsx
@@ -1,0 +1,2 @@
+export { useCartContext } from "./cart-context"
+export { useWishListContext } from "./wishlist-context"
