+++ src/context/wishlist-context.jsx
@@ -1,5 +1,5 @@
 import { useReducer,createContext ,useContext} from "react";
-import { wishReducerFunction } from "../hooks/Wishlist/wishlist-controller";
+import { wishReducerFunction } from "../hooks/Wishlist/wishlist-reducer";
 
 const WishListContext=createContext();
 export const useWishListContext=()=>useContext(WishListContext);
@@ -9,7 +9,7 @@ export const WishListProvider=({children})=>{
 const [wishListState,wishDispatch]=useReducer(wishReducerFunction,{
 wishCount:0,
 wishItems:[],
-addToishListMessage:false,
+addToWishListMessage:false,
 inWishlist:false
 })
     return <WishListContext.Provider value={{wishListState,wishDispatch}}>
