+++ src/hooks/Cart/cart.jsx
@@ -1,2 +1,2 @@
-export {reducerFunction} from "./cart-controllers"
+export {reducerFunction} from "./cart-reducer"
 export {useCartContext} from "../../context/cart-context"
